Split probabilistic models in discrete and continuous, implemented binomial distribution
